Create a package folder and move the source code into it
